--- clean.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-rm server0.txt server1.txt server2.txt server3.txt server4.txt
