+++ client.go
@@ -1,12 +1,14 @@
 package raft
-type Client struct{
+
+type Client struct {
 	servers []*Server
 }
-func (c *Client) sendCommand(serverId int,command string) bool{
+
+func (c *Client) sendCommand(serverId int, command string) bool {
 	return c.servers[serverId].cm.SubmitCommand(command)
 }
 
-func NewClient(server_list []*Server) *Client{
+func NewClient(server_list []*Server) *Client {
 	c := new(Client)
 	c.servers = server_list
 	return c
