+++ script/clean.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+rm debuglog/*.txt
