+++ script/test.sh
