+++ service_test.go
@@ -9,25 +9,25 @@ import (
 )
 
 // It is similar to a crash where server cannot send request and response to any other server
-func DisconectTwoServer(servers []*Server,id1 int,id2 int){
+func DisconectTwoServer(servers []*Server, id1 int, id2 int) {
 	servers[id1].DisconnectPeer(id2)
 	servers[id2].DisconnectPeer(id1)
 }
-func ConnectTwoServer(servers []*Server,id1 int,id2 int){
+func ConnectTwoServer(servers []*Server, id1 int, id2 int) {
 	servers[id1].ConnectToPeer(id2, servers[id2].GetListenAddr())
 	servers[id2].ConnectToPeer(id1, servers[id1].GetListenAddr())
 }
 func IsolatedServer(servers []*Server, id int, num_server int) {
 	for i := 0; i < num_server; i++ {
 		if id != i {
-			DisconectTwoServer(servers,i,id)
+			DisconectTwoServer(servers, i, id)
 		}
 	}
 }
-func RestoreIsolatedServer(servers []*Server, id int, num_server int,failures []bool) {
+func RestoreIsolatedServer(servers []*Server, id int, num_server int, failures []bool) {
 	for i := 0; i < num_server; i++ {
-		if id != i && (failures[i]==false){
-			ConnectTwoServer(servers,id,i)
+		if id != i && (failures[i] == false) {
+			ConnectTwoServer(servers, id, i)
 		}
 	}
 }
@@ -91,7 +91,7 @@ func TestElectionBasic(t *testing.T) {
 	rand.Seed(time.Now().UnixNano())
 	num_server := 5
 	servers := make([]*Server, num_server)
-	failures := make([]bool,num_server)
+	failures := make([]bool, num_server)
 	connected := make([]bool, num_server)
 	ready := make(chan interface{})
 	for i := 0; i < num_server; i++ {
@@ -140,31 +140,31 @@ func TestElectionBasic(t *testing.T) {
 			}
 
 		}
-		t.Logf("Failure machine is %d %d %d",rand_machine_1,rand_machine_2,rand_machine_3)
+		t.Logf("Failure machine is %d %d %d", rand_machine_1, rand_machine_2, rand_machine_3)
 		restore_machine_1 := true
 		restore_machine_2 := true
 		//restore_machine_3 := true
 		for {
 			elapsed := time.Now()
-			if crash_event && elapsed.Sub(start) > 3*time.Second{
+			if crash_event && elapsed.Sub(start) > 3*time.Second {
 				crash_event = false
-				failures[rand_machine_1]= true
-				failures[rand_machine_2]= true
-				failures[rand_machine_3]= true
+				failures[rand_machine_1] = true
+				failures[rand_machine_2] = true
+				failures[rand_machine_3] = true
 				IsolatedServer(servers, rand_machine_1, num_server)
 				IsolatedServer(servers, rand_machine_2, num_server)
 				IsolatedServer(servers, rand_machine_3, num_server)
 			}
 			if elapsed.Sub(start) > 5*time.Second && restore_machine_1 {
 				restore_machine_1 = false
-				t.Logf("Restore machine %d",rand_machine_1)
-				RestoreIsolatedServer(servers, rand_machine_1, num_server,failures)
+				t.Logf("Restore machine %d", rand_machine_1)
+				RestoreIsolatedServer(servers, rand_machine_1, num_server, failures)
 				failures[rand_machine_1] = false
 			}
 			if elapsed.Sub(start) > 7*time.Second && restore_machine_2 {
 				restore_machine_2 = false
-				t.Logf("Restore machine %d",rand_machine_2)
-				RestoreIsolatedServer(servers, rand_machine_2, num_server,failures)
+				t.Logf("Restore machine %d", rand_machine_2)
+				RestoreIsolatedServer(servers, rand_machine_2, num_server, failures)
 				failures[rand_machine_2] = false
 			}
 			if elapsed.Sub(start) < 11*time.Second {
@@ -183,7 +183,6 @@ func TestElectionBasic(t *testing.T) {
 	}()
 	time.Sleep(15 * time.Second)
 }
-
 
 // type Set struct {
 // 	myset map[int]struct{}
